refactor: remove the @webiny/handler-db package
With the `DbInstance` key-value store gone, `handler-db` no longer had a
reason to exist. Its two remaining pieces move to `@webiny/db-dynamodb`
(their natural home):

- `DbFeature` -> `DynamoDBCoreFeature`: registers the DynamoDB document client
  + DDB filter utilities that CMS entry filtering needs (the Feature-style
  twin of the existing `registerDynamoDBCore`).
- `dbPlugins` (deprecated test-only `context.db` glue) -> `db-dynamodb/testing`.

Consumers repointed (AWS handler, cms-es-tasks test, cms-ddb/-ddb-es test
setup), stale `@webiny/handler-db` deps dropped across the workspace (incl. the
AWS project templates), and the package deleted.
