+++ packages/api-event-handler-aws/package.json
@@ -26,10 +26,10 @@
     "@webiny/aws-sdk": "0.0.0",
     "@webiny/background-tasks-aws": "0.0.0",
     "@webiny/cognito": "0.0.0",
+    "@webiny/db-dynamodb": "0.0.0",
     "@webiny/event-handler-aws": "0.0.0",
     "@webiny/event-handler-core": "0.0.0",
-    "@webiny/handler": "0.0.0",
-    "@webiny/handler-db": "0.0.0"
+    "@webiny/handler": "0.0.0"
   },
   "peerDependencies": {
     "@webiny/di": ">=0.2.3"
