+++ packages/api-headless-cms-ddb-es/package.json
@@ -38,7 +38,6 @@
     "@webiny/db-dynamodb": "0.0.0",
     "@webiny/error": "0.0.0",
     "@webiny/feature": "0.0.0",
-    "@webiny/handler-db": "0.0.0",
     "@webiny/plugins": "0.0.0",
     "@webiny/utils": "0.0.0",
     "dataloader": "^2.2.3",
