+++ packages/api-headless-cms-ddb-es/tsconfig.json
@@ -10,7 +10,6 @@
     { "path": "../db-dynamodb" },
     { "path": "../error" },
     { "path": "../feature" },
-    { "path": "../handler-db" },
     { "path": "../plugins" },
     { "path": "../utils" },
     { "path": "../api-core" },
@@ -45,8 +44,6 @@
       "@webiny/error": ["../error/src"],
       "@webiny/feature/*": ["../feature/src/*"],
       "@webiny/feature": ["../feature/src"],
-      "@webiny/handler-db/*": ["../handler-db/src/*"],
-      "@webiny/handler-db": ["../handler-db/src"],
       "@webiny/plugins/*": ["../plugins/src/*"],
       "@webiny/plugins": ["../plugins/src"],
       "@webiny/utils/*": ["../utils/src/*"],
