+++ packages/api-headless-cms-es-tasks/package.json
@@ -28,7 +28,7 @@
   },
   "devDependencies": {
     "@webiny/build-tools": "0.0.0",
-    "@webiny/handler-db": "0.0.0",
+    "@webiny/db-dynamodb": "0.0.0",
     "@webiny/handler-graphql": "0.0.0",
     "@webiny/project-utils": "0.0.0",
     "@webiny/testing": "0.0.0",
