+++ packages/api-headless-cms-es-tasks/tsconfig.json
@@ -10,7 +10,7 @@
     { "path": "../error" },
     { "path": "../feature" },
     { "path": "../utils" },
-    { "path": "../handler-db" },
+    { "path": "../db-dynamodb" },
     { "path": "../handler-graphql" },
     { "path": "../testing" }
   ],
@@ -37,8 +37,8 @@
       "@webiny/feature": ["../feature/src"],
       "@webiny/utils/*": ["../utils/src/*"],
       "@webiny/utils": ["../utils/src"],
-      "@webiny/handler-db/*": ["../handler-db/src/*"],
-      "@webiny/handler-db": ["../handler-db/src"],
+      "@webiny/db-dynamodb/*": ["../db-dynamodb/src/*"],
+      "@webiny/db-dynamodb": ["../db-dynamodb/src"],
       "@webiny/handler-graphql/*": ["../handler-graphql/src/*"],
       "@webiny/handler-graphql": ["../handler-graphql/src"],
       "@webiny/testing/*": ["../testing/src/*"],
