+++ packages/api-headless-cms/package.json
@@ -51,7 +51,6 @@
     "@types/pluralize": "^0.0.33",
     "@webiny/build-tools": "0.0.0",
     "@webiny/db-dynamodb": "0.0.0",
-    "@webiny/handler-db": "0.0.0",
     "@webiny/project-utils": "0.0.0",
     "@webiny/sdk": "0.0.0",
     "@webiny/wcp": "0.0.0",
