+++ packages/api-mailer/package.json
@@ -34,7 +34,6 @@
     "@webiny/event-handler-core": "0.0.0",
     "@webiny/handler": "0.0.0",
     "@webiny/handler-aws": "0.0.0",
-    "@webiny/handler-db": "0.0.0",
     "@webiny/project-utils": "0.0.0",
     "graphql": "^16.14.2",
     "rimraf": "^6.1.3",
