+++ packages/api-mailer/tsconfig.json
@@ -9,8 +9,7 @@
     { "path": "../api-core" },
     { "path": "../event-handler-core" },
     { "path": "../handler" },
-    { "path": "../handler-aws" },
-    { "path": "../handler-db" }
+    { "path": "../handler-aws" }
   ],
   "compilerOptions": {
     "rootDirs": ["./src", "./__tests__"],
@@ -34,9 +33,7 @@
       "@webiny/handler/*": ["../handler/src/*"],
       "@webiny/handler": ["../handler/src"],
       "@webiny/handler-aws/*": ["../handler-aws/src/*"],
-      "@webiny/handler-aws": ["../handler-aws/src"],
-      "@webiny/handler-db/*": ["../handler-db/src/*"],
-      "@webiny/handler-db": ["../handler-db/src"]
+      "@webiny/handler-aws": ["../handler-aws/src"]
     }
   }
 }
