+++ packages/api-websockets-aws/package.json
@@ -32,7 +32,6 @@
     "@webiny/build-tools": "0.0.0",
     "@webiny/di": "^1.0.2",
     "@webiny/event-handler-core": "0.0.0",
-    "@webiny/handler-db": "0.0.0",
     "@webiny/handler-graphql": "0.0.0",
     "@webiny/project-utils": "0.0.0",
     "@webiny/wcp": "0.0.0",
