+++ packages/api-websockets-aws/tsconfig.build.json
@@ -12,7 +12,6 @@
     { "path": "../api-headless-cms/tsconfig.build.json" },
     { "path": "../api-testing/tsconfig.build.json" },
     { "path": "../event-handler-core/tsconfig.build.json" },
-    { "path": "../handler-db/tsconfig.build.json" },
     { "path": "../handler-graphql/tsconfig.build.json" },
     { "path": "../wcp/tsconfig.build.json" }
   ],
@@ -43,8 +42,6 @@
       "@webiny/api-testing": ["../api-testing/src"],
       "@webiny/event-handler-core/*": ["../event-handler-core/src/*"],
       "@webiny/event-handler-core": ["../event-handler-core/src"],
-      "@webiny/handler-db/*": ["../handler-db/src/*"],
-      "@webiny/handler-db": ["../handler-db/src"],
       "@webiny/handler-graphql/*": ["../handler-graphql/src/*"],
       "@webiny/handler-graphql": ["../handler-graphql/src"],
       "@webiny/wcp/*": ["../wcp/src/*"],
