--- packages/handler-db/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@webiny/handler-db",
-  "version": "0.0.0",
-  "type": "module",
-  "exports": {
-    ".": "./index.js",
-    "./*": "./*"
-  },
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/webiny/webiny-js.git"
-  },
-  "description": "Webiny API plugin that provides a db driver for DynamoDB.",
-  "author": "Webiny Ltd.",
-  "dependencies": {
-    "@webiny/api": "0.0.0",
-    "@webiny/aws-sdk": "0.0.0",
-    "@webiny/db": "0.0.0",
-    "@webiny/db-dynamodb": "0.0.0",
-    "@webiny/db-utils": "0.0.0",
-    "@webiny/di": "^1.0.2",
-    "@webiny/feature": "0.0.0",
-    "@webiny/handler": "0.0.0"
-  },
-  "devDependencies": {
-    "@webiny/build-tools": "0.0.0",
-    "rimraf": "^6.1.3",
-    "typescript": "^7.0.2"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "8476da73b653c89cc1474d968baf55c1b0ae0e5f",
-  "webiny": {
-    "publishFrom": "dist"
-  }
-}
