--- packages/handler-db/src/abstractions.ts
@@ -1,2 +1,0 @@
-// Re-export Bruno's DynamoDBClient abstraction + feature for convenience
-export { DynamoDBClient, DynamoDBClientFeature } from "@webiny/db-dynamodb";
