--- packages/handler-db/src/index.ts
@@ -1,3 +1,0 @@
-export { DbFeature } from "./DbFeature.js";
-export type { DbFeatureConfig } from "./DbFeature.js";
-export { DynamoDBClient, DynamoDBClientFeature } from "./abstractions.js";
