--- packages/handler-db/tsconfig.json
@@ -1,36 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src", "__tests__"],
-  "references": [
-    { "path": "../api" },
-    { "path": "../aws-sdk" },
-    { "path": "../db" },
-    { "path": "../db-dynamodb" },
-    { "path": "../db-utils" },
-    { "path": "../feature" },
-    { "path": "../handler" }
-  ],
-  "compilerOptions": {
-    "rootDirs": ["./src", "./__tests__"],
-    "outDir": "./dist",
-    "declarationDir": "./dist",
-    "paths": {
-      "~/*": ["./src/*"],
-      "~tests/*": ["./__tests__/*"],
-      "@webiny/api/*": ["../api/src/*"],
-      "@webiny/api": ["../api/src"],
-      "@webiny/aws-sdk/*": ["../aws-sdk/src/*"],
-      "@webiny/aws-sdk": ["../aws-sdk/src"],
-      "@webiny/db/*": ["../db/src/*"],
-      "@webiny/db": ["../db/src"],
-      "@webiny/db-dynamodb/*": ["../db-dynamodb/src/*"],
-      "@webiny/db-dynamodb": ["../db-dynamodb/src"],
-      "@webiny/db-utils/*": ["../db-utils/src/*"],
-      "@webiny/db-utils": ["../db-utils/src"],
-      "@webiny/feature/*": ["../feature/src/*"],
-      "@webiny/feature": ["../feature/src"],
-      "@webiny/handler/*": ["../handler/src/*"],
-      "@webiny/handler": ["../handler/src"]
-    }
-  }
-}
