--- packages/handler-db/webiny.config.js
@@ -1,8 +1,0 @@
-import { createWatchPackage, createBuildPackage } from "@webiny/build-tools";
-
-export default {
-    commands: {
-        build: createBuildPackage({ cwd: import.meta.dirname }),
-        watch: createWatchPackage({ cwd: import.meta.dirname })
-    }
-};
