+++ packages/project-aws-template/package.json
@@ -55,7 +55,6 @@
     "@webiny/aws-sdk": "0.0.0",
     "@webiny/db-dynamodb": "0.0.0",
     "@webiny/handler-aws": "0.0.0",
-    "@webiny/handler-db": "0.0.0",
     "@webiny/handler-graphql": "0.0.0",
     "@webiny/project-aws": "0.0.0",
     "@webiny/project-template-base": "0.0.0",
@@ -113,7 +112,6 @@
         "@webiny/aws-sdk",
         "@webiny/db-dynamodb",
         "@webiny/handler-aws",
-        "@webiny/handler-db",
         "@webiny/handler-graphql",
         "core-js",
         "cross-fetch",
